requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.12.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
contourpy==1.3.2
cycler==0.12.1
eval_type_backport==0.2.2
exceptiongroup==1.3.0
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.18.0
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2025.5.0
gradio==5.30.0
gradio_client==1.10.1
groovy==0.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.31.4
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.3
mdurl==0.1.2
multidict==6.4.4
narwhals==1.40.0
numpy==1.26.4
orjson==3.10.18
packaging==25.0
pandas==2.2.3
pillow==10.4.0
propcache==0.3.1
pyarrow==20.0.0
pydantic==2.11.4
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.3
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.31.0
rich==13.9.4
rpds-py==0.25.0
ruff==0.11.10
safehttpx==0.1.6
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.46.2
tabulate==0.9.0
together==1.3.14
tomlkit==0.12.0
tqdm==4.67.1
typer==0.15.4
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
websockets==11.0.3
yarl==1.20.0
